+++ vf_base/config/freemode.lua
@@ -1,5 +1,9 @@
 Config = {
 	enableTextChat = true,
 	enableIdleKick = false,	
-	idleSecondsUntilKick = 600
+	idleSecondsUntilKick = 600,
+	weapons = {
+	"weapon_knife",
+	"weapon_pistol"
+    }
 }