+++ vf_interaction/__resource.lua
@@ -1,0 +1,11 @@
+resource_manifest_version '44febabe-d386-4d18-afbe-5e627f4af937'
+
+client_scripts {
+    '@NativeUI/NativeUI.lua',
+    'config/costs.lua',
+    'config/vehicles.lua',
+    'vehicles.lua',
+	'client.lua'	
+}
+
+server_script 'server.lua'