+++ vf_interaction/config/costs.lua
@@ -1,0 +1,4 @@
+Costs = {
+	["clear_wanted_level"] = 400,
+	["repair_vehicle"] = 50,
+}