+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog    
 
-## 1.0.0
+## 1.1
+
+- Fixed random bugs.
+- Players can now use vending machines to restore health when `vf_vending` is loaded.
+- Players can manage use ATM's to deposit and whitdraw money.
+- Dropped the interaction menu from the base resource and created a separated resource `vf_interaction`.
+
+## 1.0
 
 - Added Native phone using scaleforms instead of NUI.
 - Added multiple missions.